\name{fun_par_sin_model}
\alias{fun_par_sin_model}
\title{Sin function of photosynthetic active radiation (PAR)}
\description{Creates a sin function of PAR which simulates the diel cycle. Based on data collected at Apalachicola Bay, Florida, in 2012.}
\usage{fun_par_sin_model(t)}
\arguments{
  \item{t}{time in seconds}}
\value{Returns par (W/m^2)}
\author{Created in Matlab by Maria Herrmann, converted to R code by Jill Arriola}
\keyword{ PAR }

