% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_schmidt_oxygen.R
\name{fun_schmidt_oxygen}
\alias{fun_schmidt_oxygen}
\title{Schmidt number calculation}
\usage{
fun_schmidt_oxygen(temp, salt)
}
\arguments{
\item{temp}{numeric for temperature}

\item{salt}{numeric for salinity}
}
\value{
sc (unitless)
}
\description{
Schmidt number calculation
}
\details{
Adapted from Matlab to R by Jill Arriola, 2021-03-16

THIS FUNCTION IS USED TO CALCULATE THE SCHMIDT NUMBER FOR THE AIR-SEA GAS
TRANSFER VELOCITY COMPONENT OF fwoxy.R
}
