% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwoxy.R
\name{fwoxy}
\alias{fwoxy}
\title{Main program for the forward oxygen model (fwoxy)}
\usage{
fwoxy(
  oxy_ic = oxy_ic,
  a_param = a_param,
  er_param = er_param,
  ht_in = ht_const,
  salt_in = salt_const,
  temp_in = temp_const,
  wspd_in = wspd_const
)
}
\arguments{
\item{oxy_ic}{numeric for initial oxygen concentration, mmol/m3}

\item{a_param}{numeric for light efficiency, (mmol/m3/d)/(W/m2)}

\item{er_param}{numeric for ecosystem respiration, mmol/m3/day}

\item{ht_in}{numeric for water column height, m}

\item{salt_in}{numeric for salinity, ppt}

\item{temp_in}{numeric for water temperature, C}

\item{wspd_in}{numeric for windspeed at 10 m, m/s}
}
\value{
2 plots and a data frame
}
\description{
Main program for the forward oxygen model (fwoxy)
}
\details{
OXYGEN BUDGET: TROC = GPP - ER - GASEX
Positive gas flux is from water to air

USES FUNCTIONS:
fun_par_sin_model
fun_eqb_oxygen
fun_density
fun_schmidt_oxygen
fun_gas_transfer_velocity

INPUT: oxygen concentration, time variables, initial conditions, forcing, parameters

OUTPUT: oxygen concentration, TROC, GPP, ER, GASEX
}
