% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_density.R
\name{fun_density}
\alias{fun_density}
\title{Seawater density calculation}
\usage{
fun_density(temp, salt, P)
}
\arguments{
\item{temp}{numeric for temperature}

\item{salt}{numeric for salinity}

\item{P}{numeric for pressure above atmospheric (dbar)}
}
\value{
RHO (kg/m^3)
}
\description{
Seawater density calculation
}
\details{
Adapted from Matlab code by Jill Arriola, 2021-03-16

THIS FUNCTION COMPUTES DENSITY OF SEAWATER ACCORDING TO THE
INTERNATIONALLY ACCEPTED (UNESCO) EQUATIONS.

Reference:
MILLERO AND POISSON, DEEP-SEA RESEARCH, 28A, 625-629, 1981.

THE STANDARD ERROR OF THIS EQUATION IS 3.6 X 10-3 KG/M-3.

INPUTS ARE
temp = Water Temperature (DEG. C)
salt = Salinity (PPT)
P = Pressure above atmospheric (dbar)

OUTPUT IS RHO = Density (kg/m^3)
}
