% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_eqb_oxygen.R
\name{fun_eqb_oxygen}
\alias{fun_eqb_oxygen}
\title{Oxygen saturation}
\usage{
fun_eqb_oxygen(temp, salt)
}
\arguments{
\item{temp}{numeric for temperature}

\item{salt}{numeric for salinity}
}
\value{
oxysat (mmol/m^3)
}
\description{
Oxygen saturation
}
\details{
Adapted from Matlab to R by Jill Arriola, 2021-03-16

Function to calclulate equilibrium OXYGEN concentration in seawater
from SST [C] and SSS [PSU].

Reference:
Garcia, H., and L.I. Gordon (1992), Oxygen solubility in seawater:
Better fitting equations, Limnol. Oceanogr., 376, 1307-1312.

INPUTS
temp = in situ temperature [deg C]
salt = in situ salinity [PSU]

OUTPUTS
oxysat = oxygen saturation (mmol/m^3)

USES: fun_density(temp, salt)
}
