% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_gas_transfer_velocity.R
\name{fun_gas_transfer_velocity}
\alias{fun_gas_transfer_velocity}
\title{Gas transfer velocity calculation}
\usage{
fun_gas_transfer_velocity(sc, wspd2)
}
\arguments{
\item{sc}{numeric for Schmidt number}

\item{wspd2}{numeric for windspeed}
}
\value{
kw (m/s)
}
\description{
Gas transfer velocity calculation
}
\details{
Adapted from Matlab to R by Jill Arriola, 2021-03-16

THIS FUNCTION CALCULATES THE GAS TRANSFER VELOCITY ACROSS THE AIR-SEA
INTERFACE FOR OXYGEN BASED ON THE WANNINKHOF (2014) EQUATION

Reference:
WANNINKHOF, R., LIMNOLOGY AND OCEANOGRAPHY METHODS, 12(6), 351-362, 2014.

INPUTS ARE
sc = Schmidt number (unitless)
wspd2 = Windspeed squared (m^2/s^2)
**Inputs come from fwoxy.R as vectors**

OUTPUT IS kw (m/s)
}
