% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_par_sin_model.R
\name{fun_par_sin_model}
\alias{fun_par_sin_model}
\title{Sin model of PAR}
\usage{
fun_par_sin_model(t)
}
\arguments{
\item{t}{numeric for time}
}
\value{
PAR (W/m^2)
}
\description{
Sin model of PAR
}
\details{
Adapated from Matlab to R by Jill Arriola, 2021-03-16

IDEALIZED SIN MODEL OF PHOTOSYNTHETIC ACTIVE RADIATION (PAR) IN MILITARY TIME
BASED ON DATA RETRIEVED FROM APALACHICOLA BAY, FLORIDA, 2012

Max = 400 W/m^2 at 12:00 and zero PAR between 18:00 and 6:00

INPUT: time (seconds) starting at midnight on day 1
OUTPUT: par (W/m^2)
}
